requirements.txt
requests
PyYAML
xmlrunner
